classroom attendance count
In the front, there are a whiteboard, TV, and whiteboard where the TV is represented by
a set of sliding images. Next, there are a set of seats with students’ names. The names are
selectable to indicate whether present or absent (all students are present by default). Finally,
we have the total absent count below which will be updated when the name is clicked